Add support for Delphes random seeds
You can specify a random seed in run.sh with the -s option.
Moreover, when submitting multiple files via submit.sh, each
file is automatically given a fixed seed based on its index in the
list of files submitted. This is useful to guarantee each LHE file
is run with the same random seed when running systematic variations.